+++ src/main/java/com/aureltimer/mixin/DimensionDetectionMixin.java
@@ -75,7 +75,6 @@ public class DimensionDetectionMixin {
                 }
                 
                 homeTracker.setLastHome(dimensionName);
-                LOGGER.info("🌍 DIMENSION DÉTECTÉE: '{}'", dimensionName);
             }
         }